transition usys.pl to usys.c
The whole xv6 is written in c and assembly. So why using a perl script to create the user/usys.S file.
This commit will create a new helper tools/usys.c and also moves the helper mkfs/mkfs.c to tools/mkfs.c.
The Makefile is updated to reflect this changes.