feature: Add config parameter to Makefile
Add Config parameter RAM and CPUCOUNT to the head of Makefile to call qemu with the parameter's defined in the Top of the file.
The Parameter will be used in compiler flags to define NCPU and MAXMEM parameters. This  will be used in the code to make sure the changes are reflected without any source code changes.

Added CPUSTACKSIZE to kernel/param.h to allow some experiments with the  Stacksize per CPU in future.

Added some output in kinit function to see if changes to RAM are reflected.
Remark: You have to "make clean && make qemu" after changes to the RAM Variable.