mkfs: Ignore strinop-truncation warning
mkfs uses the builtin strncpy, rather than kernel/strings.c, which warns
when the `n` is the same size as the destination buffer, as it may not
be able to add a NULL byte. This is the desired behavior for directory
entry names (which only require a NULL byte if they are shorter than
DIRSIZ), so ignore the warning to appease newer GCCs.